add home page to root directory
