+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import './App.css';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "./components/Layout";
-import Home from "./pages/homepage/Home";
+import Home from "./pages/homepage/Home2";
 import About from "./pages/About"
 import Contact from "./pages/Contact"
 import Header from "./components/Header"
@@ -12,7 +12,8 @@ function App() {
     <>
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<Home />}>
+          <Route path="/" element={<Layout />}>
+            <Route path="/" element={<Home />} />
             <Route path="home" element={<Home />} />
             <Route path="about" element={<About />} />
             <Route path="contact" element={<Contact />} />
