+++ src/products.js
@@ -92,3 +92,70 @@ export const PRODUCTS = [
     productImage: product13,
   },
 ];
+
+export const productList = [
+  {
+    id: 1,
+    productName: "IPhone",
+    price: 999.0,
+    image:
+      "https://ng.jumia.is/cms/0-1-homepage/0-0-freelinks-gray/300x240/phones-tablets_300x240.png",
+    description: "Example description",
+  },
+  {
+    id: 2,
+    productName: "MacBook Pro 2020(M1)",
+    price: 1300.0,
+    image:
+      "https://ng.jumia.is/cms/0-1-homepage/0-0-freelinks-gray/300x240/fashion_300x240.png",
+    description: "Example description",
+  },
+  {
+    id: 3,
+    productName: "Canon M50 Camera",
+    price: 1500.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/67/1602222/1.jpg?0699",
+    description: "Example description",
+  },
+  {
+    id: 4,
+    productName: "Jean shirt",
+    price: 300.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/67/1602222/1.jpg?0699",
+    description: "Example description",
+  },
+  {
+    id: 5,
+    productName: "Strip LED Light",
+    price: 10.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/26/8523401/1.jpg?6963",
+    description: "Example description",
+  },
+  {
+    id: 6,
+    productName: "Polo Shirt",
+    price: 50.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/74/2598502/1.jpg?7566",
+    description: "Example description",
+  },
+  {
+    id: 7,
+    productName: "Golf Shirt",
+    price: 1200.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/10/1033821/1.jpg?7896",
+    description: "Example description",
+  },
+  {
+    id: 8,
+    productName: "Face Cap",
+    price: 1300.0,
+    image:
+      "https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/06/7283311/1.jpg?0878",
+    description: "Example description",
+  },
+];
