--- src/products.jsx
@@ -1,119 +1,0 @@
-import one from "./images/one.png";
-import product2 from "./images/2.png";
-import product3 from "./images/3.png";
-import product4 from "./images/4.png";
-import product5 from "./images/5.png";
-import product6 from "./images/6.webp";
-import product7 from "./images/7.webp";
-import product8 from "./images/8.webp";
-
-export const PRODUCTS = [
-    {
-        id: 1,
-        productName: "IPhone",
-        price: 999.0,
-        poductImage: one,
-    },
-    {
-        id: 2,
-        productName: "MacBook Pro 2020(M1)",
-        price: 1300.0,
-        poductImage: product2,
-    },
-    {
-        id: 3,
-        productName: "Canon M50 Camera",
-        price: 1500.0,
-        poductImage: product3,
-    },
-    {
-        id: 4,
-        productName: "Jean shirt",
-        price: 300.0,
-        poductImage: product4,
-    },
-    {
-        id: 5,
-        productName: "Strip LED Light",
-        price: 10.0,
-        poductImage: product5,
-    },
-    {
-        id: 6,
-        productName: "Polo Shirt",
-        price: 50.0,
-        poductImage: product6,
-    },
-    {
-        id: 7,
-        productName: "Golf Shirt",
-        price: 120.0,
-        poductImage: product7,
-    },
-    {
-        id: 8,
-        productName: "Face Cap",
-        price: 13.0,
-        poductImage: product8,
-    },
-]
-
-
-export const productList = [
-    {
-        id: 1,
-        productName: "IPhone",
-        price: 999.0,
-        image: 'https://ng.jumia.is/cms/0-1-homepage/0-0-freelinks-gray/300x240/phones-tablets_300x240.png',
-        description: 'Example description',
-    },
-    {
-        id: 2,
-        productName: "MacBook Pro 2020(M1)",
-        price: 1300.0,
-        image: 'https://ng.jumia.is/cms/0-1-homepage/0-0-freelinks-gray/300x240/fashion_300x240.png',
-        description: 'Example description',
-    },
-    {
-        id: 3,
-        productName: "Canon M50 Camera",
-        price: 1500.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/67/1602222/1.jpg?0699',
-        description: 'Example description',
-    },
-    {
-        id: 4,
-        productName: "Jean shirt",
-        price: 300.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/67/1602222/1.jpg?0699',
-        description: 'Example description',
-    },
-    {
-        id: 5,
-        productName: "Strip LED Light",
-        price: 10.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/26/8523401/1.jpg?6963',
-        description: 'Example description',
-    },
-    {
-        id: 6,
-        productName: "Polo Shirt",
-        price: 50.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/74/2598502/1.jpg?7566',
-        description: 'Example description',
-    },
-    {
-        id: 7,
-        productName: "Golf Shirt",
-        price: 1200.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/10/1033821/1.jpg?7896',
-        description: 'Example description',
-    },
-    {
-        id: 8,
-        productName: "Face Cap",
-        price: 1300.0,
-        image: 'https://ng.jumia.is/unsafe/fit-in/300x300/filters:fill(white)/product/06/7283311/1.jpg?0878',
-        description: 'Example description',
-    },
-]
